charms/keystone-k8s/requirements.txt
charmhelpers
jinja2
kubernetes
ops
ops-lib-mysql
python-keystoneclient
git+https://opendev.org/openstack/charm-ops-openstack#egg=ops_openstack
